+++ sdk/runtime/constants.go
@@ -1,9 +1,0 @@
-package runtime
-
-const (
-	SubjectResult    = "sys.job.result"
-	SubjectHeartbeat = "sys.heartbeat"
-	SubjectProgress  = "sys.job.progress"
-	SubjectCancel    = "sys.job.cancel"
-	DefaultProtocolVersion = 1
-)